requirements.txt
# LinkedIn API
linkedin-api==2.0.0

# Gestion des secrets
python-dotenv==1.0.0

# Intelligence Artificielle - Modèles
openai==1.3.0          # Pour OpenAI GPT
anthropic==0.5.0       # Pour Claude
llama-cpp-python==0.2.6 # Pour Llama local
requests==2.31.0       # Pour Ollama API

# Base de données
sqlalchemy==2.0.23

# Interface utilisateur
streamlit==1.29.0

# Tests
pytest==7.4.3

# Utilitaires
python-dateutil==2.8.2
schedule==1.2.1 